load commands as subapp in app.py
